Fixed max speed EEPROM load bug
The value for the motor max speeds was being loaded from EEPROM into an
incorrect variable type, thereby causing it to not get properly passed
to the motor objects.

+++ Firmware/Motion_Engine/OM_EEPROM.ino
@@ -158,7 +158,7 @@ void eepromRestore() {
     long tempStop   = 0;
     long tempEnd    = 0;
     float tempAccel  = 0;
-    float tempMaxSpd = 0;
+    unsigned int tempMaxSpd = 0;
 
 
     for (int i = 0; i < MOTOR_COUNT; i++){
